+++ crates/ink/codegen/src/generator/metadata.rs
@@ -365,7 +365,6 @@ impl Metadata<'_> {
                 .hash(#hash)
                 .timestamp(#timestamp)
                 .block_number(#block_number)
-                .max_event_topics(MAX_EVENT_TOPICS)
                 .native_to_eth_ratio(NATIVE_TO_ETH_RATIO)
                 .static_buffer_size(#buffer_size_const)
                 .done()