+++ crates/ink/tests/ui/contract/pass/config-custom-env.rs
@@ -4,7 +4,6 @@
 pub struct CustomEnv;
 
 impl ink_env::Environment for CustomEnv {
-    const MAX_EVENT_TOPICS: usize = 3;
     const NATIVE_TO_ETH_RATIO: u32 = 100_000_000;
     type AccountId = [u8; 32];
     type Balance = u64;