+++ crates/metadata/src/specs.rs
@@ -280,7 +280,7 @@ where
             "only one default message is allowed"
         );
 
-        let max_topics = self.spec.environment.max_event_topics;
+        const MAX_TOPICS: usize = 4;
         let events_exceeding_max_topics_limit = self
             .spec
             .events
@@ -289,7 +289,7 @@ where
                 let signature_topic = if e.signature_topic.is_some() { 1 } else { 0 };
                 let topics_count =
                     signature_topic + e.args.iter().filter(|a| a.indexed).count();
-                if topics_count > max_topics {
+                if topics_count > MAX_TOPICS {
                     Some(format!(
                         "`{}::{}` ({} topics)",
                         e.module_path, e.label, topics_count
@@ -301,7 +301,7 @@ where
             .collect::<Vec<_>>();
         assert!(
             events_exceeding_max_topics_limit.is_empty(),
-            "maximum of {max_topics} event topics exceeded: {}",
+            "maximum of {MAX_TOPICS} event topics exceeded: {}",
             events_exceeding_max_topics_limit.join(", ")
         );
 
@@ -655,8 +655,6 @@ mod state {
     pub struct Timestamp;
     /// Type state for the `BlockNumber` type of the environment.
     pub struct BlockNumber;
-    /// Type state for the max number of topics specified in the environment.
-    pub struct MaxEventTopics;
     /// Type state for the native to eth ratio specified in the environment.
     pub struct NativeToEthRatio;
     /// Type state for the size of the static buffer configured via environment variable.`
@@ -1550,7 +1548,6 @@ where
     hash: TypeSpec<F>,
     timestamp: TypeSpec<F>,
     block_number: TypeSpec<F>,
-    max_event_topics: usize,
     native_to_eth_ratio: u32,
     static_buffer_size: usize,
 }
@@ -1567,7 +1564,6 @@ where
             hash: Default::default(),
             timestamp: Default::default(),
             block_number: Default::default(),
-            max_event_topics: Default::default(),
             native_to_eth_ratio: Default::default(),
             static_buffer_size: Default::default(),
         }
@@ -1584,7 +1580,6 @@ impl IntoPortable for EnvironmentSpec {
             hash: self.hash.into_portable(registry),
             timestamp: self.timestamp.into_portable(registry),
             block_number: self.block_number.into_portable(registry),
-            max_event_topics: self.max_event_topics,
             native_to_eth_ratio: self.native_to_eth_ratio,
             static_buffer_size: self.static_buffer_size,
         }
@@ -1616,10 +1611,6 @@ where
     pub fn block_number(&self) -> &TypeSpec<F> {
         &self.block_number
     }
-    /// Returns the `MAX_EVENT_TOPICS` value of the environment.
-    pub fn max_event_topics(&self) -> usize {
-        self.max_event_topics
-    }
     /// Returns the `NATIVE_TO_ETH_RATIO` value of the environment.
     pub fn native_to_eth_ratio(&self) -> u32 {
         self.native_to_eth_ratio
@@ -1640,7 +1631,6 @@ where
         Missing<state::Hash>,
         Missing<state::Timestamp>,
         Missing<state::BlockNumber>,
-        Missing<state::MaxEventTopics>,
         Missing<state::NativeToEthRatio>,
         Missing<state::BufferSize>,
     > {
@@ -1654,18 +1644,18 @@ where
 /// An environment specification builder.
 #[allow(clippy::type_complexity)]
 #[must_use]
-pub struct EnvironmentSpecBuilder<F, A, B, H, T, BN, M, NTER, BS>
+pub struct EnvironmentSpecBuilder<F, A, B, H, T, BN, NTER, BS>
 where
     F: Form,
     TypeSpec<F>: Default,
     EnvironmentSpec<F>: Default,
 {
     spec: EnvironmentSpec<F>,
-    marker: PhantomData<fn() -> (A, B, H, T, BN, M, NTER, BS)>,
+    marker: PhantomData<fn() -> (A, B, H, T, BN, NTER, BS)>,
 }
 
-impl<F, B, H, T, BN, M, NTER, BS>
-    EnvironmentSpecBuilder<F, Missing<state::AccountId>, B, H, T, BN, M, NTER, BS>
+impl<F, B, H, T, BN, NTER, BS>
+    EnvironmentSpecBuilder<F, Missing<state::AccountId>, B, H, T, BN, NTER, BS>
 where
     F: Form,
     TypeSpec<F>: Default,
@@ -1675,7 +1665,7 @@ where
     pub fn account_id(
         self,
         account_id: TypeSpec<F>,
-    ) -> EnvironmentSpecBuilder<F, state::AccountId, B, H, T, BN, M, NTER, BS> {
+    ) -> EnvironmentSpecBuilder<F, state::AccountId, B, H, T, BN, NTER, BS> {
         EnvironmentSpecBuilder {
             spec: EnvironmentSpec {
                 account_id,
@@ -1686,8 +1676,8 @@ where
     }
 }
 
-impl<F, A, H, T, BN, M, NTER, BS>
-    EnvironmentSpecBuilder<F, A, Missing<state::Balance>, H, T, BN, M, NTER, BS>
+impl<F, A, H, T, BN, NTER, BS>
+    EnvironmentSpecBuilder<F, A, Missing<state::Balance>, H, T, BN, NTER, BS>
 where
     F: Form,
     TypeSpec<F>: Default,
@@ -1697,7 +1687,7 @@ where
     pub fn balance(
         self,
         balance: TypeSpec<F>,
-    ) -> EnvironmentSpecBuilder<F, A, state::Balance, H, T, BN, M, NTER, BS> {
+    ) -> EnvironmentSpecBuilder<F, A, state::Balance, H, T, BN, NTER, BS> {
         EnvironmentSpecBuilder {
             spec: EnvironmentSpec {
                 balance,
@@ -1708,8 +1698,8 @@ where
     }
 }
 
-impl<F, A, B, T, BN, M, NTER, BS>
-    EnvironmentSpecBuilder<F, A, B, Missing<state::Hash>, T, BN, M, NTER, BS>
+impl<F, A, B, T, BN, NTER, BS>
+    EnvironmentSpecBuilder<F, A, B, Missing<state::Hash>, T, BN, NTER, BS>
 where
     F: Form,
     TypeSpec<F>: Default,
@@ -1719,7 +1709,7 @@ where
     pub fn hash(
         self,
         hash: TypeSpec<F>,
-    ) -> EnvironmentSpecBuilder<F, A, B, state::Hash, T, BN, M, NTER, BS> {
+    ) -> EnvironmentSpecBuilder<F, A, B, state::Hash, T, BN, NTER, BS> {
         EnvironmentSpecBuilder {
             spec: EnvironmentSpec { hash, ..self.spec },
             marker: PhantomData,
@@ -1727,8 +1717,8 @@ where
     }
 }
 
-impl<F, A, B, H, BN, M, NTER, BS>
-    EnvironmentSpecBuilder<F, A, B, H, Missing<state::Timestamp>, BN, M, NTER, BS>
+impl<F, A, B, H, BN, NTER, BS>
+    EnvironmentSpecBuilder<F, A, B, H, Missing<state::Timestamp>, BN, NTER, BS>
 where
     F: Form,
     TypeSpec<F>: Default,
@@ -1738,7 +1728,7 @@ where
     pub fn timestamp(
         self,
         timestamp: TypeSpec<F>,
-    ) -> EnvironmentSpecBuilder<F, A, B, H, state::Timestamp, BN, M, NTER, BS> {
+    ) -> EnvironmentSpecBuilder<F, A, B, H, state::Timestamp, BN, NTER, BS> {
         EnvironmentSpecBuilder {
             spec: EnvironmentSpec {
                 timestamp,
@@ -1749,8 +1739,8 @@ where
     }
 }
 
-impl<F, A, B, H, T, M, NTER, BS>
-    EnvironmentSpecBuilder<F, A, B, H, T, Missing<state::BlockNumber>, M, NTER, BS>
+impl<F, A, B, H, T, NTER, BS>
+    EnvironmentSpecBuilder<F, A, B, H, T, Missing<state::BlockNumber>, NTER, BS>
 where
     F: Form,
     TypeSpec<F>: Default,
@@ -1760,7 +1750,7 @@ where
     pub fn block_number(
         self,
         block_number: TypeSpec<F>,
-    ) -> EnvironmentSpecBuilder<F, A, B, H, T, state::BlockNumber, M, NTER, BS> {
+    ) -> EnvironmentSpecBuilder<F, A, B, H, T, state::BlockNumber, NTER, BS> {
         EnvironmentSpecBuilder {
             spec: EnvironmentSpec {
                 block_number,
@@ -1771,30 +1761,8 @@ where
     }
 }
 
-impl<F, A, B, H, T, BN, NTER, BS>
-    EnvironmentSpecBuilder<F, A, B, H, T, BN, Missing<state::MaxEventTopics>, NTER, BS>
-where
-    F: Form,
-    TypeSpec<F>: Default,
-    EnvironmentSpec<F>: Default,
-{
-    /// Sets the `MAX_EVENT_TOPICS` value of the environment.
-    pub fn max_event_topics(
-        self,
-        max_event_topics: usize,
-    ) -> EnvironmentSpecBuilder<F, A, B, H, T, BN, state::MaxEventTopics, NTER, BS> {
-        EnvironmentSpecBuilder {
-            spec: EnvironmentSpec {
-                max_event_topics,
-                ..self.spec
-            },
-            marker: PhantomData,
-        }
-    }
-}
-
-impl<F, A, B, H, T, BN, M, BS>
-    EnvironmentSpecBuilder<F, A, B, H, T, BN, M, Missing<state::NativeToEthRatio>, BS>
+impl<F, A, B, H, T, BN, BS>
+    EnvironmentSpecBuilder<F, A, B, H, T, BN, Missing<state::NativeToEthRatio>, BS>
 where
     F: Form,
     TypeSpec<F>: Default,
@@ -1804,7 +1772,7 @@ where
     pub fn native_to_eth_ratio(
         self,
         native_to_eth_ratio: u32,
-    ) -> EnvironmentSpecBuilder<F, A, B, H, T, BN, M, state::NativeToEthRatio, BS> {
+    ) -> EnvironmentSpecBuilder<F, A, B, H, T, BN, state::NativeToEthRatio, BS> {
         EnvironmentSpecBuilder {
             spec: EnvironmentSpec {
                 native_to_eth_ratio,
@@ -1815,8 +1783,8 @@ where
     }
 }
 
-impl<F, A, B, H, T, BN, M, NTER>
-    EnvironmentSpecBuilder<F, A, B, H, T, BN, M, NTER, Missing<state::BufferSize>>
+impl<F, A, B, H, T, BN, NTER>
+    EnvironmentSpecBuilder<F, A, B, H, T, BN, NTER, Missing<state::BufferSize>>
 where
     F: Form,
     TypeSpec<F>: Default,
@@ -1826,7 +1794,7 @@ where
     pub fn static_buffer_size(
         self,
         static_buffer_size: usize,
-    ) -> EnvironmentSpecBuilder<F, A, B, H, T, BN, M, NTER, state::BufferSize> {
+    ) -> EnvironmentSpecBuilder<F, A, B, H, T, BN, NTER, state::BufferSize> {
         EnvironmentSpecBuilder {
             spec: EnvironmentSpec {
                 static_buffer_size,
@@ -1845,7 +1813,6 @@ impl<F>
         state::Hash,
         state::Timestamp,
         state::BlockNumber,
-        state::MaxEventTopics,
         state::NativeToEthRatio,
         state::BufferSize,
     >